+++ packages/react-ai/src/hooks/useAIGeneration.tsx
@@ -1,4 +1,5 @@
-import { DataState, useDataState } from '@aws-amplify/ui-react-core';
+import * as React from 'react';
+import { DataState } from '@aws-amplify/ui-react-core';
 import { V6Client } from '@aws-amplify/api-graphql';
 import { getSchema } from '../types';
 
@@ -9,7 +10,7 @@ export interface UseAIGenerationHookWrapper<
   useAIGeneration: <U extends Key>(
     routeName: U
   ) => [
-    Awaited<DataState<Schema[U]['returnType']>>,
+    Awaited<GenerationState<Schema[U]['returnType']>>,
     (input: Schema[U]['args']) => void,
   ];
 }
@@ -20,7 +21,7 @@ export type UseAIGenerationHook<
 > = (
   routeName: Key
 ) => [
-  Awaited<DataState<Schema[Key]['returnType']>>,
+  Awaited<GenerationState<Schema[Key]['returnType']>>,
   (input: Schema[Key]['args']) => void,
 ];
 
@@ -37,6 +38,24 @@ interface GraphQLFormattedError {
   };
 }
 
+type SingularReturnValue<T> = {
+  data: T | null;
+  errors?: GraphQLFormattedError[];
+};
+
+type GenerationState<T> = Omit<DataState<T>, 'message'> & {
+  messages?: GraphQLFormattedError[];
+};
+
+// default state
+const INITIAL_STATE = {
+  hasError: false,
+  isLoading: false,
+  messages: undefined,
+};
+const LOADING_STATE = { hasError: false, isLoading: true, messages: undefined };
+const ERROR_STATE = { hasError: true, isLoading: false };
+
 export function createUseAIGeneration<
   Client extends Record<'generations' | 'conversations', Record<string, any>>,
   Schema extends getSchema<Client>,
@@ -45,31 +64,43 @@ export function createUseAIGeneration<
     Key extends keyof AIGenerationClient<Schema>['generations'],
   >(
     routeName: Key
-  ) => {
-    const handleGenerate = (
-      client.generations as AIGenerationClient<Schema>['generations']
-    )[routeName];
+  ): [
+    state: GenerationState<Schema[Key]['returnType']>,
+    handleAction: (input: Schema[Key]['args']) => void,
+  ] => {
+    const [dataState, setDataState] = React.useState<
+      GenerationState<Schema[Key]['returnType']>
+    >(() => ({
+      ...INITIAL_STATE,
+      data: undefined,
+    }));
 
-    const updateAIGenerationStateAction = async (
-      _prev: Schema[Key]['returnType'],
-      input: Schema[Key]['args']
-    ): Promise<
-      Schema[Key]['returnType'] & { graphqlErrors?: GraphQLFormattedError[] }
-    > => {
-      const result = await handleGenerate(input);
+    const handleGeneration = React.useCallback(
+      async (input: Schema[Key]['args']) => {
+        setDataState(({ data }) => ({ ...LOADING_STATE, data }));
 
-      // handleGenerate returns a Promised wrapper around Schema[Key]['returnType'] which includes data, errors, and clientExtensions
-      // The type of data is Schema[Key]['returnType'] which useDataState also wraps in a data return
-      // TODO: follow up with how to type handleGenerate to properly return the promise wrapper shape
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-      const data = (result as any).data as Schema[Key]['returnType'];
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
-      const graphqlErrors = (result as any).errors;
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-assignment
-      return { ...data, ...(graphqlErrors ? { graphqlErrors } : {}) };
-    };
+        const result = await (
+          client.generations as AIGenerationClient<Schema>['generations']
+        )[routeName](input);
 
-    return useDataState(updateAIGenerationStateAction, {});
+        const { data, errors } = result as SingularReturnValue<
+          Schema[Key]['returnType']
+        >;
+
+        if (errors) {
+          setDataState({
+            ...ERROR_STATE,
+            data,
+            messages: errors,
+          });
+        } else {
+          setDataState({ ...INITIAL_STATE, data });
+        }
+      },
+      [routeName]
+    );
+
+    return [dataState, handleGeneration];
   };
 
   return useAIGeneration;