+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/ui-react
 
+## 6.5.4
+
+### Patch Changes
+
+- Updated dependencies [[`ac7cb271a`](https://github.com/aws-amplify/amplify-ui/commit/ac7cb271aff895e643fb5dc927030df9245b7c5b)]:
+  - @aws-amplify/ui@6.6.4
+  - @aws-amplify/ui-react-core@3.0.28
+
 ## 6.5.3
 
 ### Patch Changes